Do not remove OpenID roles on logout
There is no guarantee that logout will be run anyway, and removing the roles
is actively harmful, because it prevents an editor from attributing posts to
an author after logout.

OpenID roles are reset to their current values on login, so there is no risk
in this arrangement that a user will receive increased privileges due to
their roles remaining while they are logged out.